+++ lib/page_magic/element/method_observer.rb
@@ -1,0 +1,13 @@
+module PageMagic
+  class Element
+    module MethodObserver
+      def singleton_method_added(arg)
+        @singleton_methods_added = true unless arg == :singleton_method_added
+      end
+
+      def singleton_methods_added?
+        @singleton_methods_added == true
+      end
+    end
+  end
+end
