+++ lib/page_magic/element/selector_methods.rb
@@ -1,0 +1,10 @@
+module PageMagic
+  class Element
+    module SelectorMethods
+      def selector(selector = nil)
+        return @selector unless selector
+        @selector = selector
+      end
+    end
+  end
+end
