--- lib/page_magic/selector_methods.rb
@@ -1,8 +1,0 @@
-module PageMagic
-  module SelectorMethods
-    def selector(selector = nil)
-      return @selector unless selector
-      @selector = selector
-    end
-  end
-end
